tasks/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
python-dotenv==0.20.0
drf-yasg==1.21.3